Deviner la géographie communale

Détecter le millésime du code officiel géographique d'une table de données communales.

COG_akinator(vecteur_codgeo, donnees_insee = T, liste_complete = T)

Arguments

vecteur_codgeo
correspond à un vecteur de codes Insee communaux pour lequel on cherche à déterminer le millésime de la géographie.

donnees_insee
vaut TRUE si les données manipulées sont produites par l'Insee. En effet, quelques rares modifications communales (la défusion des communes Loisey et Culey au 1er janvier 2014 par exemple) ont été prises en compte dans les bases de données communales de l'Insee plus tard que la date officielle. Pour tous les COG officiels datant d'avant 2008, seules les tables de passage Insee sont disponibles dans ce package.

liste_complete
vaut TRUE si le vecteur de communes du paramètre *vecteur_codgeo* liste l'ensemble des communes françaises et FALSE s'il en liste seulement un extrait.

Details

Le code officiel géographique de référence du package est actuellement celui au 01/01/2017. Les données communales devront être dans ce COG pour être agrégées en niveaux supra-communaux (fonction nivsupra).

Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 1968, 1975, 1982, 1990, 1999, 2007 à 2017.

Les dates de référence des codes officiels géographiques utilisés dans COGugaison sont les suivantes :

COG 1968 : à partir du 01/03/1968

COG 1975 : à partir du 20/02/1975

COG 1982 : à partir du 04/03/1982

COG 1990 : à partir du 05/03/1990

COG 1999 : à partir du 08/03/1999

Pour tous les autres COG : à partir du 01/01 de chaque année

Les différences entre les tables de passage Insee et non Insee sont les suivantes :

1982-03-03 (pris en compte par l'Insee seulement après le 04/03/1982): Flaignes-Havys (08169) est un rassemblement de Flaignes-Havys (08169), Havys (08221) [fusion simple].

2014-01-01 (pris en compte par l'Insee seulement au 01/01/2015) : Loisey (55298) s'est séparée en Loisey (55298), Culey (55138) [rétablissement].

1990-02-01 (pris en compte par l'Insee seulement après le 05/03/1990) : Le code commune de Oudon passe de 14624 à 14697 [changement de code dû à un changement de chef-lieu].

1981-09-28 (pris en compte par l'Insee dès le 20/02/1975) : Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir des parcelles d'Incarville (27351), de Léry (27365) , de Porte-Joie (27471) , de Poses (27474) , de Saint-Étienne-du-Vauvray (27537), de Saint-Pierre-du-Vauvray (27598), de Tournedos-sur-Seine (27651) et du Vaudreuil (27528) [création]. Cette situation étant complexe, nous avons pour le moment considéré que Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir de parcelles du Vaudreuil (27528) uniquement.

En 1968, les 4 communes qui auraient dû d'après le COG être codées 2B044,2B076,2B151 et 2A325 sont codées 20044,20076,20151 et 20325 dans les données Insee.

References

historique des géographies communales (Insee)

tables d'appartenance des communes aux différents niveaux géographiques (Insee)

See also

changement_COG_varNum, changement_COG_typo,changement_COG_typo_details, enlever_PLM, modification_Corse, modifications_communales,nivsupra

Examples

## Exemple 1
# Ici, nous cherchons le millésime du code officiel géographique (COG) utilisé dans la table de données communales exemple_popcom.
COG_akinator(vecteur_codgeo=exemple_popcom[,1],donnees_insee=T, liste_complete=T)

#> [1] "COG2014"