Modifier les codes Insee de l'ancienne commune de l'Oudon (Calvados, 14)

Modifier les codes Insee de l'ancienne commune de l'Oudon (Calvados, 14). En effet, la commune de l'Oudon a toujours été codée "14697" puis a changé de code devenant "14472" à partir du 07/01/2014 suite à un transfert de chef lieu. Cette commune a depuis disparu puisqu'elle est devenue commune déléguée au sein de Saint-Pierre-en-Auge (14654).

modification_Oudon(
  table_entree,
  codgeo_entree = colnames(table_entree)[1],
  vecteur_entree = is.vector(table_entree),
  donnees_insee_entree = TRUE,
  donnees_insee_sortie = TRUE,
  COG = as.numeric(substr(COG_akinator(vecteur_codgeo = table_entree[, codgeo_entree],
    donnees_insee = donnees_insee_entree), 4, 7))
)

Arguments

table_entree

correspond à la table (ou le vecteur, cf. paramètre vecteur_entree) à transformer de manière à modifier le code de la commune de l'Oudon.

codgeo_entree

est une chaîne de caractères qui indique le nom de la variable contenant les codes communes Insee. Par défaut, il s'agit du nom de la première colonne de table_entree.

vecteur_entree

vaut TRUE si table_entree est un simple vecteur.

donnees_insee_entree

vaut TRUE si les données en entrée sont des bases de données de l'Insee

donnees_insee_sortie

vaut TRUE si les données souhaitées en sorties doivent respecter le même COG que les données Insee

COG

indique l'année de COG de la table considérée. (exemple 1968). Par défaut renvoie l'année retournée par la fonction COG_akinator

Details

Le code officiel géographique le plus récent du package est actuellement celui au 01/01/2023.

Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 1968, 1975, 1982, 1990, 1999, 2008 à 2023.

Les dates de référence des codes officiels géographiques utilisés dans COGugaison sont les suivantes :

COG 1968 : à partir du 01/03/1968

COG 1975 : à partir du 20/02/1975

COG 1982 : à partir du 04/03/1982

COG 1990 : à partir du 05/03/1990

COG 1999 : à partir du 08/03/1999

Pour tous les autres COG : à partir du 01/01 de chaque année

Les différences entre les tables de passage Insee et non Insee sont les suivantes :

1982-03-03 (pris en compte par l'Insee seulement après le 04/03/1982): Flaignes-Havys (08169) est un rassemblement de Flaignes-Havys (08169), Havys (08221) [fusion simple].

2014-01-01 (pris en compte par l'Insee seulement au 01/01/2015) : Loisey (55298) s'est séparée en Loisey (55298), Culey (55138) [rétablissement].

1990-02-01 (pris en compte par l'Insee seulement après le 05/03/1990) : Le code commune de Oudon passe de 14624 à 14697 [changement de code dû à un changement de chef-lieu].

2014-01-07 (pris en compte par l'Insee dès le 01/01/2016) : Tôtes est rattachée à Notre-Dame-de-Fresnay qui devient L'Oudon (changement de code de l'Oudon de 14697 à 14472) [transfert de chef-lieu].

1981-09-28 (pris en compte par l'Insee dès le 20/02/1975) : Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir des parcelles d'Incarville (27351), de Léry (27365) , de Porte-Joie (27471) , de Poses (27474) , de Saint-Étienne-du-Vauvray (27537), de Saint-Pierre-du-Vauvray (27598), de Tournedos-sur-Seine (27651) et du Vaudreuil (27528) [création]. Cette situation étant complexe, nous avons pour le moment considéré que Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir de parcelles du Vaudreuil (27528) uniquement.

En 1968, les 4 communes qui auraient dû d'après le COG être codées 2B044,2B076,2B151 et 2A325 sont codées 20044,20076,20151 et 20325 dans les données Insee.

References

historique des géographies communales (Insee)

tables d'appartenance des communes aux différents niveaux géographiques (Insee)

See also

Examples

## Exemple 1
# Ici, nous allons remplacer les codes communes de l'Oudon dans une table de l'Insee (en COG 2014 mais que nous considérons en COG 2015 uniquement pour cet exemple).
head(exemple_popcom[which(exemple_popcom$CODGEO%in%c("14472","14697")),])
#>      CODGEO  LIBGEO P12_POP SUPERF    typoA typoB
#> 5159  14697 L'Oudon    1551     55 CLASSE 1     D
exemple_popcom_oudon <- modification_Oudon(table_entree=exemple_popcom,donnees_insee_entree=T,donnees_insee_sortie=F,COG=2015)
head(exemple_popcom_oudon[which(exemple_popcom$CODGEO%in%c("14472","14697")),])
#>      CODGEO  LIBGEO P12_POP SUPERF    typoA typoB
#> 5159  14472 L'Oudon    1551     55 CLASSE 1     D