Agréger des données à des échelons supra-communaux

Permettre d'agréger les tables de données communales à de nombreux échelons supra-communaux administratifs (EPCI, arrondissements, cantons-villes, départements, régions) ou d'étude (bassins de vie, zones d'emploi, unités urbaines, aires urbaines). La fonction `nivsupra` peut également s'appliquer à des tables de flux (non agrégées par codes communes) grâce à l'option agregation = FALSE. La table en entrée est alors conservée comme telle avec une nouvelle colonne qui correspond au niveau supracommunal du code commune considéré.

nivsupra(
  table_entree,
  codgeo_entree = colnames(table_entree)[1],
  COG = annee_ref,
  var_num = colnames(table_entree)[sapply(table_entree, is.numeric)],
  nivsupra,
  nivsupra_nom = ifelse(agregation == TRUE, nivsupra, paste0(nivsupra, "_",
    codgeo_entree)),
  agregation = T
)

Arguments

table_entree

correspond à la table (ou le vecteur, cf. paramètre vecteur_entree) à transformer de manière à modifier les codes des communes corses.

codgeo_entree

est une chaîne de caractères qui indique le nom de la variable contenant les codes communes Insee. Par défaut, il s'agit du nom de la première colonne de table_entree.

COG

indique l'année de COG de la table communale considérée. (exemple 2014). Années possibles : de 2008 à 2020. Par défaut, vaut 2020.

var_num

est un vecteur de chaînes de caractères qui indique les noms des variables numériques à convertir. Par défaut, il s'agit de l'ensemble des variables de types numériques dans table_entree.

nivsupra

est une chaîne de caractères qui indique le nom du niveau supra-communale souhaité. Pour les années récentes : - "DEP" : départements - "REG" : régions - "EPCI" : EPCI au 01/01/20XX - "ARR" : arrondissements au 01/01/20XX - "CV" : cantons-villes au 01/01/20XX - "ZE2010" : zones d'emploi 2010 (avant 2020) - "ZE2020" : zones d'emploi 2020 (après 2020) - "UU2010" : unités urbaines 2010 (avant 2020) - "UU2020" : unités urbaines 2020 (après 2020) - "AU2010" : aires urbaines 2010 (avant 2020) - "AAV2020" : aires d'attraction des villes 2020 (après 2020) - "BV2012" : bassins de vie 2012 Pour les COG plus anciens, regarder au cas par cas selon les années. Par exemple, str(table_supracom_2008)

nivsupra_nom

indique le nom à donner au niveau supra-communal dans la table de sortie. Il faut par défaut la chaîne de caractère contenue dans nivsupra si agregation = T et la concaténation de nivsupra et codgeo_entree séparée d'un "_" si agregation = F.

agregation

vaut TRUE si la table souhaitée doit sommer toutes les lignes qui concernent une même commune et FALSE si l'on souhaite volontairement conserver les doublons dans les codes commune (dans les tables de flux par exemple). Si agregation = F, les variables de type caractère sont alors conservées comme telles ou dupliquées en cas de défusion et les variables numériques sommées en cas de fusion ou réparties proportionnellement à la population de chaque commune en cas de défusion.

Details

Le code officiel géographique le plus récent du package est actuellement celui au 01/01/2023.

Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 1968, 1975, 1982, 1990, 1999, 2008 à 2023.

A partir de 2020, de nombreux zonages ont été modifiés suite à la refonte des zonages de l'Insee. Les ZE2010 deviennent les ZE2020, les UU2010 deviennent les UU2020, les AU2010 deviennent les AAV2020.

Les dates de référence des codes officiels géographiques utilisés dans COGugaison sont les suivantes :

COG 1968 : à partir du 01/03/1968

COG 1975 : à partir du 20/02/1975

COG 1982 : à partir du 04/03/1982

COG 1990 : à partir du 05/03/1990

COG 1999 : à partir du 08/03/1999

Pour tous les autres COG : à partir du 01/01 de chaque année

Les différences entre les tables de passage Insee et non Insee sont les suivantes :

1982-03-03 (pris en compte par l'Insee seulement après le 04/03/1982): Flaignes-Havys (08169) est un rassemblement de Flaignes-Havys (08169), Havys (08221) [fusion simple].

2014-01-01 (pris en compte par l'Insee seulement au 01/01/2015) : Loisey (55298) s'est séparée en Loisey (55298), Culey (55138) [rétablissement].

1990-02-01 (pris en compte par l'Insee seulement après le 05/03/1990) : Le code commune de Oudon passe de 14624 à 14697 [changement de code dû à un changement de chef-lieu].

2014-01-07 (pris en compte par l'Insee dès le 01/01/2016) : Tôtes est rattachée à Notre-Dame-de-Fresnay qui devient L'Oudon (changement de code de l'Oudon de 14697 à 14472) [transfert de chef-lieu].

1981-09-28 (pris en compte par l'Insee dès le 20/02/1975) : Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir des parcelles d'Incarville (27351), de Léry (27365) , de Porte-Joie (27471) , de Poses (27474) , de Saint-Étienne-du-Vauvray (27537), de Saint-Pierre-du-Vauvray (27598), de Tournedos-sur-Seine (27651) et du Vaudreuil (27528) [création]. Cette situation étant complexe, nous avons pour le moment considéré que Vaudreuil-Ex-Ensemble Urbain (27701) est créée à partir de parcelles du Vaudreuil (27528) uniquement.

En 1968, les 4 communes qui auraient dû d'après le COG être codées 2B044,2B076,2B151 et 2A325 sont codées 20044,20076,20151 et 20325 dans les données Insee.

References

historique des géographies communales (Insee)

tables d'appartenance des communes aux différents niveaux géographiques (Insee)

See also

Examples

## Exemple 1
# Ici, apres avoir transforme les donnees en geographie communale au 01/01/2017, nous agregeons la population et la superficie des communes a l'echelon geographique des zones d'emploi afin d'obtenir une table des densites de population par zone d'emploi.
exemple_popcom_COG2017_num <- changement_COG_varNum(table_entree=exemple_popcom,annees=c(2014:2017),agregation=T,libgeo=T,donnees_insee=T)
exemple_popcom_ZE2010 <- nivsupra(table_entree=exemple_popcom_COG2017_num,COG=2017,nivsupra="ZE2010",agregation=T)
exemple_popcom_ZE2010$densite <- exemple_popcom_ZE2010$P12_POP / exemple_popcom_ZE2010$SUPERF
head(exemple_popcom_ZE2010)
#>   ZE2010                         LIBGEO P12_POP SUPERF  densite
#> 1   0050                 Mont-de-Marsan  160717   4956 32.42877
#> 2   0051                        Alençon  122263   2405 50.83701
#> 3   0052                Cosne - Clamecy   75052   2553 29.39757
#> 4   0053                          Mâcon  143358   1543 92.90862
#> 5   0054               Nogent-le-Rotrou   48665   1227 39.66178
#> 6   0055 La Vallée de la Bresle - Vimeu   99095   1134 87.38536
## Exemple 2
# Ici, on ajoute les colonnes ZE2010_COMMUNE et ZE2010_DCLT a la table exemple de flux domicile-travail.
exemple_flux_COG2017 <- changement_COG_varNum(table_entree=exemple_flux,annees=c(2014:2017),codgeo_entree="COMMUNE",agregation=F,libgeo=F,donnees_insee=T)
exemple_flux_COG2017 <- changement_COG_varNum(table_entree=exemple_flux_COG2017,annees=c(2014:2017),codgeo_entree="DCLT",agregation=F,libgeo=F,donnees_insee=T)
exemple_flux_COG2017_etZE <- nivsupra(table_entree=exemple_flux_COG2017,codgeo_entree="COMMUNE",nivsupra="ZE2010",COG=2017,agregation=F)
exemple_flux_COG2017_etZE <- nivsupra(table_entree=exemple_flux_COG2017_etZE,codgeo_entree="DCLT",nivsupra="ZE2010",COG=2017,agregation=F)
head(exemple_flux_COG2017_etZE)
#>   COMMUNE  DCLT CS1 AGEREVQ   IPONDI ZE2010_COMMUNE ZE2010_nom_COMMUNE
#> 1   61001 61001   4      25 2.856389           0051            Alençon
#> 2   61001 61001   3      55 3.040441           0051            Alençon
#> 3   61213 53185   2      45 4.142349           0051            Alençon
#> 4   61464 61001   5      35 4.059227           0051            Alençon
#> 5   61491 61500   6      15 5.578419           0051            Alençon
#> 6   61117 61117   2      45 3.945892           0051            Alençon
#>   ZE2010_DCLT ZE2010_nom_DCLT
#> 1        0051         Alençon
#> 2        0051         Alençon
#> 3        0051         Alençon
#> 4        0051         Alençon
#> 5        0051         Alençon
#> 6        0051         Alençon