Créer une table de communes fusionnées appartenant à des classes différentes selon une typologie

Isoler dans une table les communes fusionnées entre deux géographies communales qui appartiennent à des classes différentes selon une typologie.

changement_COG_typo_details(table_entree, annees,
  codgeo_entree = colnames(table_entree)[1], typo, donnees_insee = T)

Arguments

table_entree
correspond à la table à transformer en une autre géographie

annees
est un vecteur qui liste l'ensemble des années qui séparent le code officiel géographique de départ et d'arrivée. Par exemple c(1968:1985). Le package rend possible l'utilisation de tables de passages d'une année de COG vers une année antiérieure (par exemple c(2016:2014)).

codgeo_entree
est une chaîne de caractères qui indique le nom de la variable contenant les codes Insee communaux. Par défaut, il s'agit du nom de la première colonne de table_entree.

typo
est une chaîne de caractères qui indique le nom de la typologie à convertir.

donnees_insee
vaut TRUE si les données manipulées sont produites par l'Insee. En effet, quelques rares modifications communales (la défusion des communes Loisey et Culey au 1er janvier 2014 par exemple) ont été prises en compte dans les bases de données communales de l'Insee plus tard que la date officielle. Pour tous les COG officiels datant d'avant 2008, seules les tables de passage Insee sont disponibles dans ce package.

Details

Le code officiel géographique de référence du package est actuellement celui au 01/01/2017. Les données communales devront être dans ce COG pour être agrégées en niveaux supra-communaux (fonction nivsupra).

Les autres codes officiels géographiques utilisés dans COGugaison sont les suivants :

COG 1968 : à partir du 01/03/1968 (donnees_insee=T obligatoirement)

COG 1975 : à partir du 20/02/1975 (donnees_insee=T obligatoirement)

COG 1982 : à partir du 04/03/1982 (donnees_insee=T obligatoirement)

COG 1990 : à partir du 05/03/1990 (donnees_insee=T obligatoirement)

COG 1999 : à partir du 08/03/1999 (donnees_insee=T obligatoirement)

COG 2008 à 2017 : à partir du 01/01 de chaque année (donnees_insee=T ou F)

Les différences entre les tables de passage Insee et non Insee sont les suivantes :

1982-03-03 (pris en compte par l'Insee seulement après le 04/03/1982): Flaignes-Havys (08169) est un rassemblement de Flaignes-Havys (08169), Havys (08221) [fusion simple].

2014-01-01 (pris en compte par l'Insee seulement au 01/01/2015) : Loisey (55298) s'est séparée en Loisey (55298), Culey (55138) [rétablissement].

1990-02-01 (pris en compte par l'Insee seulement après le 05/03/1990) : Le code commune de Oudon passe de 14624 à 14697 [changement de code dû à un changement de chef-lieu].

References

historique des géographies communales (Insee)

tables d'appartenance des communes aux différents niveaux géographiques (Insee)

See also

changement_COG_varNum, changement_COG_typo, COG_akinator, enlever_PLM, modification_Corse, modifications_communales,nivsupra

Examples

## Exemple 1
# Ici nous allons transformer les deux typologies (typoA et typoB) de la table exemple_pop en géographie communale au 1er janvier 2017 (au lieu de 2014).
# L'hypothèse de classement en cas de fusion de communes (*methode_fusion*) choisie est celle d'une classe spécifique (*methode_difference*, classe appelée *mot_difference*="differents") aux regroupements de plusieurs communes de classes différentes. Les autres hypothèses possibles auraient pu être l'hypothèse du maximum de population *methode_max_pop* ou de classe absorbante *methode_classe_absorbante*.
exemple_popcom_COG2017_typo <- changement_COG_typo(table_entree=exemple_popcom[,-2],annees=c(2014:2017),methode_fusion="methode_difference",typos=c("typoA","typoB"),mot_difference = "differents",libgeo="LIB",donnees_insee=T)
head(exemple_popcom_COG2017_typo)

#>   CODGEO                     LIB    typoA typoB
#> 1  01001 L'Abergement-Clémenciat CLASSE 2     F
#> 2  01002   L'Abergement-de-Varey CLASSE 3     D
#> 3  01004       Ambérieu-en-Bugey CLASSE 1     B
#> 4  01005     Ambérieux-en-Dombes CLASSE 3     F
#> 5  01006                 Ambléon CLASSE 4     E
#> 6  01007                Ambronay CLASSE 5     A

# Nous allons maintenant isoler dans une table les communes fusionnées appartenant à des classes différentes, ici selon la typologie "typoA" entre 2014 et 2015, 2015 et 2016 et 2016 et 2017.
details_exemple_popcom_COG2017_typo <- changement_COG_typo_details(table_entree=exemple_popcom[,-2],annees=c(2014:2017),typo="typoA", donnees_insee=T)
head(details_exemple_popcom_COG2017_typo[["2014_2015"]])

#>   cod2014              lib2014 cod2015                     lib2015    typoA
#> 1   14178               Corbon   14474 Notre-Dame-d'Estrées-Corbon CLASSE 2
#> 2   14474 Notre-Dame-d'Estrées   14474 Notre-Dame-d'Estrées-Corbon CLASSE 4
#> 3   25034        Auxon-Dessous   25035                  Les Auxons CLASSE 1
#> 4   25035         Auxon-Dessus   25035                  Les Auxons CLASSE 3
#> 5   38024           Badinières   38152           Eclose-Badinières CLASSE 4
#> 6   38152               Eclose   38152           Eclose-Badinières CLASSE 3

head(details_exemple_popcom_COG2017_typo[["2015_2016"]])

#>   cod2015             lib2015 cod2016            lib2016    typoA
#> 1   01015           Arbignieu   01015    Arboys en Bugey CLASSE 3
#> 2   01340          Saint-Bois   01015    Arboys en Bugey CLASSE 2
#> 3   01119           Corcelles   01080 Champdor-Corcelles CLASSE 1
#> 4   01080            Champdor   01080 Champdor-Corcelles CLASSE 2
#> 5   01176 Le Grand-Abergement   01187      Haut Valromey CLASSE 2
#> 6   01409             Songieu   01187      Haut Valromey CLASSE 5

head(details_exemple_popcom_COG2017_typo[["2016_2017"]])

#>   cod2016             lib2016 cod2017          lib2017    typoA
#> 1   01095 Chavannes-sur-Suran   01095 Nivigne et Suran CLASSE 5
#> 2   01172           Germagnat   01095 Nivigne et Suran CLASSE 1
#> 3   01098         Chazey-Bons   01098      Chazey-Bons CLASSE 3
#> 4   01316              Pugieu   01098      Chazey-Bons CLASSE 4
#> 5   03168             Meaulne   03168   Meaulne-Vitray CLASSE 2
#> 6   03318              Vitray   03168   Meaulne-Vitray CLASSE 5